requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
wave
tempfile
base64
io
os